deleting files w the fs package

+++ 13_FileUpload-download-and-pagination/controllers/admin.js
@@ -1,4 +1,5 @@
 const mongoose = require("mongoose");
+const fileHelper = require("../util/file");
 
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
@@ -140,6 +141,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = description;
 
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
@@ -186,15 +188,21 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const { productId } = req.body;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
-    .then(() => {
-      console.log("SUCCESSFULLY DELETED!!");
-      res.redirect("/admin/products");
-    })
-    .catch((err) => {
-      // res.redirect("/500");
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(productId).then((product) => {
+    if (!product) {
+      return next(new Error("Product not found"));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: productId, userId: req.user._id })
+      .then(() => {
+        console.log("SUCCESSFULLY DELETED!!");
+        res.redirect("/admin/products");
+      })
+      .catch((err) => {
+        // res.redirect("/500");
+        const error = new Error(err);
+        error.httpStatusCode = 500;
+        return next(error);
+      });
+  });
 };
